Enforce 1-second record spacing for all Activities.
Records are typically recorded at 1-second intervals, but the odd
timestamp might be off by a second or two for various reasons.
This resulted in multiple records with the same offset index,
which prevented indexing based on timestamp alone. There were
other issues as well.
This fix naively assumes 1-second record spacing, which is how
my files always are.
In the future, I would like to investigate the effects of pauses
in the file. If paused records are not recorded, it would not make
sense to enforce 1-second spacing.